requirement.txt
Django==4.2.21
mysqlclient==2.1.1
djangorestframework==3.15.2